Example of configuring builtin plugin.
