+++ pkg/types/kustomization.go
@@ -5,7 +5,6 @@
 package types
 
 import (
-	"sigs.k8s.io/kustomize/v3/pkg/gvk"
 	"sigs.k8s.io/kustomize/v3/pkg/image"
 )
 
@@ -19,13 +18,6 @@ const (
 type TypeMeta struct {
 	Kind       string `json:"kind,omitempty" yaml:"kind,omitempty"`
 	APIVersion string `json:"apiVersion,omitempty" yaml:"apiversion,omitempty"`
-}
-
-// ObjectMeta partially copies apimachinery/pkg/apis/meta/v1.ObjectMeta
-// No need for a direct dependence; the fields are stable.
-type ObjectMeta struct {
-	Name      string `json:"name,omitempty" yaml:"name,omitempty"`
-	Namespace string `json:"namespace,omitempty" yaml:"namespace,omitempty"`
 }
 
 // Kustomization holds the information needed to generate customized k8s api resources.
@@ -193,26 +185,6 @@ func (k *Kustomization) EnforceFields() []string {
 	return errs
 }
 
-// GeneratorArgs contains arguments common to generators.
-type GeneratorArgs struct {
-	// Namespace for the configmap, optional
-	Namespace string `json:"namespace,omitempty" yaml:"namespace,omitempty"`
-
-	// Name - actually the partial name - of the generated resource.
-	// The full name ends up being something like
-	// NamePrefix + this.Name + hash(content of generated resource).
-	Name string `json:"name,omitempty" yaml:"name,omitempty"`
-
-	// Behavior of generated resource, must be one of:
-	//   'create': create a new one
-	//   'replace': replace the existing one
-	//   'merge': merge with the existing one
-	Behavior string `json:"behavior,omitempty" yaml:"behavior,omitempty"`
-
-	// DataSources for the generator.
-	DataSources `json:",inline,omitempty" yaml:",inline,omitempty"`
-}
-
 // PluginConfig holds plugin configuration.
 type PluginConfig struct {
 	// DirectoryPath is an absolute path to a
@@ -225,75 +197,10 @@ type PluginConfig struct {
 	Enabled bool
 }
 
-// ConfigMapArgs contains the metadata of how to generate a configmap.
-type ConfigMapArgs struct {
-	// GeneratorArgs for the configmap.
-	GeneratorArgs `json:",inline,omitempty" yaml:",inline,omitempty"`
-}
-
 // Pair is a key value pair.
 type Pair struct {
 	Key   string
 	Value string
-}
-
-// SecretArgs contains the metadata of how to generate a secret.
-type SecretArgs struct {
-	// GeneratorArgs for the secret.
-	GeneratorArgs `json:",inline,omitempty" yaml:",inline,omitempty"`
-
-	// Type of the secret.
-	//
-	// This is the same field as the secret type field in v1/Secret:
-	// It can be "Opaque" (default), or "kubernetes.io/tls".
-	//
-	// If type is "kubernetes.io/tls", then "literals" or "files" must have exactly two
-	// keys: "tls.key" and "tls.crt"
-	Type string `json:"type,omitempty" yaml:"type,omitempty"`
-}
-
-// DataSources contains some generic sources for configmaps.
-type DataSources struct {
-	// LiteralSources is a list of literal
-	// pair sources. Each literal source should
-	// be a key and literal value, e.g. `key=value`
-	LiteralSources []string `json:"literals,omitempty" yaml:"literals,omitempty"`
-
-	// FileSources is a list of file "sources" to
-	// use in creating a list of key, value pairs.
-	// A source takes the form:  [{key}=]{path}
-	// If the "key=" part is missing, the key is the
-	// path's basename. If they "key=" part is present,
-	// it becomes the key (replacing the basename).
-	// In either case, the value is the file contents.
-	// Specifying a directory will iterate each named
-	// file in the directory whose basename is a
-	// valid configmap key.
-	FileSources []string `json:"files,omitempty" yaml:"files,omitempty"`
-
-	// EnvSources is a list of file paths.
-	// The contents of each file should be one
-	// key=value pair per line, e.g. a Docker
-	// or npm ".env" file or a ".ini" file
-	// (wikipedia.org/wiki/INI_file)
-	EnvSources []string `json:"envs,omitempty" yaml:"envs,omitempty"`
-
-	// Deprecated.  Use EnvSources instead.
-	EnvSource string `json:"env,omitempty" yaml:"env,omitempty"`
-}
-
-// GeneratorOptions modify behavior of all ConfigMap and Secret generators.
-type GeneratorOptions struct {
-	// Labels to add to all generated resources.
-	Labels map[string]string `json:"labels,omitempty" yaml:"labels,omitempty"`
-
-	// Annotations to add to all generated resources.
-	Annotations map[string]string `json:"annotations,omitempty" yaml:"annotations,omitempty"`
-
-	// DisableNameSuffixHash if true disables the default behavior of adding a
-	// suffix to the names of generated resources that is a hash of the
-	// resource contents.
-	DisableNameSuffixHash bool `json:"disableNameSuffixHash,omitempty" yaml:"disableNameSuffixHash,omitempty"`
 }
 
 type PluginType string
@@ -336,30 +243,6 @@ type PatchJson6902 struct {
 	Patch string `json:"patch,omitempty" yaml:"patch,omitempty"`
 }
 
-// PatchTarget represents the kubernetes object that the patch is applied to
-type PatchTarget struct {
-	gvk.Gvk   `json:",inline,omitempty" yaml:",inline,omitempty"`
-	Namespace string `json:"namespace,omitempty" yaml:"namespace,omitempty"`
-	Name      string `json:"name" yaml:"name"`
-}
-
-// PatchStrategicMerge represents a relative path to a
-// stategic merge patch with the format
-// https://github.com/kubernetes/community/blob/master/contributors/devel/strategic-merge-patch.md
-type PatchStrategicMerge string
-
-// Replica specifies a modification to a replica config.
-// The number of replicas of a resource whose name matches will be set to count.
-// This struct is used by the ReplicaCountTransform, and is meant to supplement
-// the existing patch functionality with a simpler syntax for replica configuration.
-type Replica struct {
-	// The name of the resource to change the replica count
-	Name string `json:"name,omitempty" yaml:"name,omitempty"`
-
-	// The number of replicas required.
-	Count int64 `json:"count,omitempty" yaml:"count,omitempty"`
-}
-
 // Patch represent either a Strategic Merge Patch or a JSON patch
 // and its targets.
 // The content of the patch can either be from a file
@@ -373,23 +256,4 @@ type Patch struct {
 
 	// Target points to the resources that the patch is applied to
 	Target *Selector `json:"target,omitempty" yaml:"target,omitempty"`
-}
-
-// Selector specifies a set of resources.
-// Any resource that matches intersection of all conditions
-// is included in this set.
-type Selector struct {
-	gvk.Gvk   `json:",inline,omitempty" yaml:",inline,omitempty"`
-	Namespace string `json:"namespace,omitempty" yaml:"namespace,omitempty"`
-	Name      string `json:"name,omitempty" yaml:"name,omitempty"`
-
-	// AnnotationSelector is a string that follows the label selection expression
-	// https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#api
-	// It matches with the resource annotations.
-	AnnotationSelector string `json:"annotationSelector,omitempty" yaml:"annotationSelector,omitempty"`
-
-	// LabelSelector is a string that follows the label selection expression
-	// https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#api
-	// It matches with the resource labels.
-	LabelSelector string `json:"labelSelector,omitempty" yaml:"labelSelector,omitempty"`
 }
